Refactor maintenance workorder task logic into separate inheritance file
